Fetch more objects, support fetcher configuration
This commit extends the type of objects collected by the fetcher. Now
all of the object types from the design spec are supported by the
project.

The ability to configure the fetcher is also added via usage of a
Kubernetes ConfigMap. A sample YAML file is included. Apart from the
ability to enable/disable specific object types, the configuration can
limit the namespaces to collect from and/or ignore specific namespaces.