Send fetched objects to an API, add unit tests
This commit modifies the fetcher to send the objects collected from
Kubernetes to an API endpoint rather than simply printing them to
standard output.

The objects will be serialized to JSON and compressed with the zstd
algorithm (although if compression fails, the data will be sent
uncompressed, servers are expected to parse according to the
Content-Encoding header).

Unit tests are also added in this commit.