Add sample files for Kubernetes RBAC
This commit introduces two new sample files for deployment of the
collector: a Cluster Role and a Cluster Role Binding. Before this
addition, the README recommended using the default roles created by
Kubernetes, but this may not provide all permissions that are necessary
by default.

All sample files are moved to the new "data" directory.

Also, the Local Development Quick Start section of the READMe is
modified to use the "regular" `kubectl` implementation rather than the
one bundled with `minikube`, as that implementation does not support
service account management as is necessary.