Use OAuth to authn with app-server; switch compression
Instead of using an API key, the collector will now use an OAuth
authentication flow using an access key and secret key supplied by the
app-server. The collector will now accept an INFRALIGHT_ACCESS_KEY and
INFRALIGHT_SECRET_KEY environment variables instead of the previous
INFRALIGH_API_KEY. The secret template is updated to include these new
credentials.

Requests to the app-server with collected objects will no longer be
compressed with zstd, but with gzip instead. This is because adding zstd
support in NodeJS and express.js is too difficult.