Take cluster ID as input, fix sending of object kinds
This commit performs two modifications to the collector:

1. The collector now requires a cluster ID to be provided as input
(either as a command line argument or an environment variable named
`CLUSTER_ID`). This is required because there is no way to get a unique
cluster identifier from the Kubernetes API server.

2. The collector will now manually add a Kind to each object, due to an
issue with how the Kubernetes client encodes objects to JSON, which
causes the kind to be missing.