Add a Helm Chart for the collector (#2)
This commit introduces a Helm Chart for the collector, which allows a
quick and easy way to install the collector, alleviating the need to
manually create dependencies such as service accounts, roles,
configmaps, etc.

The Helm Chart is configurable, allowing customers to modify options
such as enabling collection of secrets easily.

The chart will eventually be deployed to Artifact Hub, but currently can
only be installed by cloning the project repository and running `helm`
directly on the chart directory.

The README file is updated to reflect these changes, and both the quick
start sections for production usage and local development are modified
to use the chart instead of performing manual steps.

The collector itself is also modified to read configuration file from a
directory (/etc/config in production, .config in development) rather
than loading a ConfigMap from the Kubernetes API Server. This is more
conformant with ConfigMap usage, which are expected to be mounted into
pods. The Helm Chart does this automatically.

The collector's Dockerfile is updated to also copy root certificate
files, so that the image accepts Infralight's TLS certificate when
connecting to the App Server.

Resolves: INFL-94