Collect deployed Helm releases; refactor (#3)
* Modify send API endpoint

The API endpoint to which Kubernetes objects are to be sent has been
changed from `POST /k8scollector` to `PUT /integrations/k8s/:clusterId`.

* Collect deployed Helm releases; refactor

This commit introduces support for collecting deployed releases via the
Helm SDK. Helm is a widely used package manager for Kubernetes (which
can also be used to install this project).

The collector image will now attempt to use the Helm SDK by default, but
this is optional in case Helm is not installed or used on the Kubernetes
cluster.

A refactor of the codebase is made to allow for multiple sources of
Kubernetes-related data:

- The collector package now defines a `DataCollector` interface, to be
  implemented by collectors of various sources. The collector object now
  accepts a list of objects implementing this interface, and will use
  them all to collect data.

- The code collecting data from the Kubernetes API server is moved to
  its own package (collector/k8s) and modified to properly implement the
  `DataCollector` interface.

- A new package, called collector/helm, also implements the interface,
  collecting deployed releases from the Helm package manager.

- The collector configuration object is moved to a separate package
  called collector/config. This is necessary because it is shared to the
  data collectors.

Documentation is added to all public assets of the code.

It should be noted that this is a breaking change. The collector's API
is modified, and also the structure of the data sent to the Infralight
App Server.

* Add API version to Kubernetes resources fetching (#4)

See https://github.com/kubernetes/kubernetes/pull/63972 for more details

* Add Kubernetes supported resources fetcher. (#6)

As part of the collector process, we wish to fetch the supported resources and their version and namespacing definitions.

Co-authored-by: Liav Yona <liav@infralight.co>