Add support for Helm charts installed via ArgoCD
This commit modifies the collector to add support for Helm charts that
were installed via ArgoCD. Since ArgoCD doesn't use Helm directly, but
rather unpack a chart's templates and applies them, "release" objects
are not created and thus not listable via `helm list`.

To workaround this, the collector will look for Kubernetes resources of
kind "Application" (an ArgoCD CRD), and generate Chart objects from
them (assuming their source is Helm).

The collector's chart template for the cluster role is modified to
include collection of ArgoCD Applications as a configuration option.

A `-dry-run` flag is added to the collector. When provided, the
collector doesn't contact Firefly servers, neither to perform
authentication nor to send collected data, which will instead be printed
in JSON to standard output.